Display root nodes of GCM/RARC in GUI, fix bugs importing RARCs
This fixes being unable to import RARCs where the root node had files as direct children of it (e.g. in Super Mario Sunshine or Super Mario Galaxy) as opposed to always having directories as children of the root node like Wind Waker.
This also allows adding new files to the root of a GCM.