Yaz0 tab: Add compression options, make compression more faithful
Added a slider to control the speed of compression vs how small the file gets compressed, as well as a checkbox to pad the compressed file (fixes #5). Also improved the accuracy of compression so that files generated by GCFT should now exactly match the original Nintendo compressed files, as long as you use the correct settings.