Fix preview model not loading if opened from a different tab
If OpenGL wasn't instantly initialized after calling self.show(), J3DUltra wouldn't be able to load a model, and would return None.

This could happen if opening a model via "Open J3D" in the RARC tab, as the tab would not be switched to until after the model was loaded, causing the OpenGL initialization to be delayed.

Also added error checking so that an error will be displayed to the user if J3DUltra returned None.