RARC: Fix contents not showing after . and ..
In standard RARCs, . and .. are placed after all of the other files and folders in each folder. But in RARCs produced by unofficial tools, . and .. can potentially be placed at the beginning of each folder. This cause the GUI code not showing anything that appears after . and .. when those two pseudo-directories are hidden from the GUI.